+++ pkg/controller/seldondeployment/seldondeployment_controller.go
@@ -30,7 +30,6 @@ import (
 	autoscaling "k8s.io/api/autoscaling/v2beta1"
 	corev1 "k8s.io/api/core/v1"
 	"k8s.io/apimachinery/pkg/api/errors"
-	"k8s.io/apimachinery/pkg/api/resource"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/runtime"
 	"k8s.io/apimachinery/pkg/types"
@@ -200,193 +199,6 @@ func getEngineEnvAnnotations(mlDep *machinelearningv1alpha2.SeldonDeployment) []
 	return envVars
 }
 
-// Create the Container for the service orchestrator.
-func createEngineContainer(mlDep *machinelearningv1alpha2.SeldonDeployment, p *machinelearningv1alpha2.PredictorSpec, engine_http_port, engine_grpc_port int) (*corev1.Container, error) {
-	// Get engine user
-	var engineUser int64 = -1
-	if engineUserEnv, ok := os.LookupEnv("ENGINE_CONTAINER_USER"); ok {
-		user, err := strconv.Atoi(engineUserEnv)
-		if err != nil {
-			return nil, err
-		} else {
-			engineUser = int64(user)
-		}
-	}
-	// get predictor as base64 encoded json
-	pCopy := p.DeepCopy()
-	// Set traffic to zero to ensure this doesn't cause a diff in the resulting  deployment created
-	pCopy.Traffic = 0
-	predictorB64, err := getEngineVarJson(pCopy)
-	if err != nil {
-		return nil, err
-	}
-
-	//get annotation for java opts or default
-	javaOpts := getAnnotation(mlDep, machinelearningv1alpha2.ANNOTATION_JAVA_OPTS, "-Dcom.sun.management.jmxremote.rmi.port=9090 -Dcom.sun.management.jmxremote -Dcom.sun.management.jmxremote.port=9090 -Dcom.sun.management.jmxremote.ssl=false -Dcom.sun.management.jmxremote.authenticate=false -Dcom.sun.management.jmxremote.local.only=false -Djava.rmi.server.hostname=127.0.0.1")
-
-	//Engine resources
-	engineResources := p.SvcOrchSpec.Resources
-	if engineResources == nil {
-		cpuQuantity, _ := resource.ParseQuantity("0.1")
-		engineResources = &corev1.ResourceRequirements{
-			Requests: map[corev1.ResourceName]resource.Quantity{
-				corev1.ResourceCPU: cpuQuantity,
-			},
-		}
-	}
-
-	c := corev1.Container{
-		Name:                     "seldon-container-engine",
-		Image:                    getEnv("ENGINE_CONTAINER_IMAGE_AND_VERSION", "seldonio/engine:0.2.7-SNAPSHOT"),
-		ImagePullPolicy:          corev1.PullPolicy(getEnv("ENGINE_CONTAINER_IMAGE_PULL_POLICY", "IfNotPresent")),
-		TerminationMessagePath:   "/dev/termination-log",
-		TerminationMessagePolicy: corev1.TerminationMessageReadFile,
-		VolumeMounts: []corev1.VolumeMount{
-			{
-				Name:      machinelearningv1alpha2.PODINFO_VOLUME_NAME,
-				MountPath: machinelearningv1alpha2.PODINFO_VOLUME_PATH,
-			},
-		},
-		Env: []corev1.EnvVar{
-			{Name: "ENGINE_PREDICTOR", Value: predictorB64},
-			{Name: "DEPLOYMENT_NAME", Value: mlDep.Spec.Name},
-			{Name: "DEPLOYMENT_NAMESPACE", Value: mlDep.ObjectMeta.Namespace},
-			{Name: "ENGINE_SERVER_PORT", Value: strconv.Itoa(engine_http_port)},
-			{Name: "ENGINE_SERVER_GRPC_PORT", Value: strconv.Itoa(engine_grpc_port)},
-			{Name: "JAVA_OPTS", Value: javaOpts},
-		},
-		Ports: []corev1.ContainerPort{
-			{ContainerPort: int32(engine_http_port), Protocol: corev1.ProtocolTCP},
-			{ContainerPort: int32(engine_grpc_port), Protocol: corev1.ProtocolTCP},
-			{ContainerPort: 8082, Name: "admin", Protocol: corev1.ProtocolTCP},
-			{ContainerPort: 9090, Name: "jmx", Protocol: corev1.ProtocolTCP},
-		},
-		ReadinessProbe: &corev1.Probe{Handler: corev1.Handler{HTTPGet: &corev1.HTTPGetAction{Port: intstr.FromString("admin"), Path: "/ready", Scheme: corev1.URISchemeHTTP}},
-			InitialDelaySeconds: 20,
-			PeriodSeconds:       1,
-			FailureThreshold:    1,
-			SuccessThreshold:    1,
-			TimeoutSeconds:      2},
-		LivenessProbe: &corev1.Probe{Handler: corev1.Handler{HTTPGet: &corev1.HTTPGetAction{Port: intstr.FromString("admin"), Path: "/live", Scheme: corev1.URISchemeHTTP}},
-			InitialDelaySeconds: 20,
-			PeriodSeconds:       5,
-			FailureThreshold:    7,
-			SuccessThreshold:    1,
-			TimeoutSeconds:      2},
-		Lifecycle: &corev1.Lifecycle{
-			PreStop: &corev1.Handler{
-				Exec: &corev1.ExecAction{Command: []string{"/bin/sh", "-c", "curl 127.0.0.1:" + strconv.Itoa(engine_http_port) + "/pause; /bin/sleep 10"}},
-			},
-		},
-		Resources: *engineResources,
-	}
-
-	if engineUser != -1 {
-		var procMount = corev1.DefaultProcMount
-		c.SecurityContext = &corev1.SecurityContext{RunAsUser: &engineUser, ProcMount: &procMount}
-	}
-
-	// Environment vars if specified
-	svcOrchEnvMap := make(map[string]string)
-	if p.SvcOrchSpec.Env != nil {
-		for _, env := range p.SvcOrchSpec.Env {
-			c.Env = append(c.Env, *env)
-			svcOrchEnvMap[env.Name] = env.Value
-		}
-	}
-
-	engineEnvVarsFromAnnotations := getEngineEnvAnnotations(mlDep)
-	for _, envVar := range engineEnvVarsFromAnnotations {
-		//don't add env vars that are already present in svcOrchSpec
-		if _, ok := svcOrchEnvMap[envVar.Name]; ok {
-			//present so don't try to overwrite
-		} else {
-			c.Env = append(c.Env, envVar)
-			// now put in map so we know it's there
-			svcOrchEnvMap[envVar.Name] = envVar.Value
-		}
-
-	}
-
-	if _, ok := svcOrchEnvMap["SELDON_LOG_MESSAGES_EXTERNALLY"]; ok {
-		//this env var is set already so no need to set a default
-	} else {
-		c.Env = append(c.Env, corev1.EnvVar{Name: "SELDON_LOG_MESSAGES_EXTERNALLY", Value: getEnv("ENGINE_LOG_MESSAGES_EXTERNALLY", "false")})
-	}
-	return &c, nil
-}
-
-// Create the service orchestrator.
-func createEngineDeployment(mlDep *machinelearningv1alpha2.SeldonDeployment, p *machinelearningv1alpha2.PredictorSpec, seldonId string, engine_http_port, engine_grpc_port int) (*appsv1.Deployment, error) {
-
-	var terminationGracePeriodSecs = int64(20)
-	var defaultMode = corev1.DownwardAPIVolumeSourceDefaultMode
-
-	depName := machinelearningv1alpha2.GetServiceOrchestratorName(mlDep, p)
-
-	con, err := createEngineContainer(mlDep, p, engine_http_port, engine_grpc_port)
-	if err != nil {
-		return nil, err
-	}
-	deploy := &appsv1.Deployment{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:        depName,
-			Namespace:   getNamespace(mlDep),
-			Labels:      map[string]string{machinelearningv1alpha2.Label_svc_orch: "true", machinelearningv1alpha2.Label_seldon_app: seldonId, machinelearningv1alpha2.Label_seldon_id: seldonId, "app": depName, "version": "v1", "fluentd": "true"},
-			Annotations: mlDep.Spec.Annotations,
-		},
-		Spec: appsv1.DeploymentSpec{
-			Selector: &metav1.LabelSelector{
-				MatchLabels: map[string]string{machinelearningv1alpha2.Label_seldon_app: seldonId, machinelearningv1alpha2.Label_seldon_id: seldonId},
-			},
-			Template: corev1.PodTemplateSpec{
-				ObjectMeta: metav1.ObjectMeta{
-					Labels: map[string]string{machinelearningv1alpha2.Label_seldon_app: seldonId, machinelearningv1alpha2.Label_seldon_id: seldonId, "app": depName},
-					Annotations: map[string]string{
-						"prometheus.io/path":   getEnv("ENGINE_PROMETHEUS_PATH", "/prometheus"),
-						"prometheus.io/port":   strconv.Itoa(engine_http_port),
-						"prometheus.io/scrape": "true",
-					},
-				},
-				Spec: corev1.PodSpec{
-					Containers: []corev1.Container{
-						*con,
-					},
-					TerminationGracePeriodSeconds: &terminationGracePeriodSecs,
-					DNSPolicy:                     corev1.DNSClusterFirst,
-					SchedulerName:                 "default-scheduler",
-					SecurityContext:               &corev1.PodSecurityContext{},
-					Volumes: []corev1.Volume{
-						{Name: machinelearningv1alpha2.PODINFO_VOLUME_NAME, VolumeSource: corev1.VolumeSource{DownwardAPI: &corev1.DownwardAPIVolumeSource{Items: []corev1.DownwardAPIVolumeFile{
-							{Path: "annotations", FieldRef: &corev1.ObjectFieldSelector{FieldPath: "metadata.annotations", APIVersion: "v1"}},
-						}, DefaultMode: &defaultMode}}},
-					},
-					RestartPolicy: corev1.RestartPolicyAlways,
-				},
-			},
-			Strategy: appsv1.DeploymentStrategy{RollingUpdate: &appsv1.RollingUpdateDeployment{MaxUnavailable: &intstr.IntOrString{StrVal: "10%"}}},
-			Replicas: &p.Replicas,
-		},
-	}
-
-	// Add a particular service account rather than default for the engine
-	if svcAccount, ok := os.LookupEnv("ENGINE_CONTAINER_SERVICE_ACCOUNT_NAME"); ok {
-		deploy.Spec.Template.Spec.ServiceAccountName = svcAccount
-		deploy.Spec.Template.Spec.DeprecatedServiceAccount = svcAccount
-	} else {
-		deploy.Spec.Template.Spec.ServiceAccountName = "default"
-		deploy.Spec.Template.Spec.DeprecatedServiceAccount = "default"
-	}
-
-	// add predictor labels
-	for k, v := range p.Labels {
-		deploy.ObjectMeta.Labels[k] = v
-		deploy.Spec.Template.ObjectMeta.Labels[k] = v
-	}
-
-	return deploy, nil
-}
-
 func createHpa(podSpec *machinelearningv1alpha2.SeldonPodSpec, deploymentName string, seldonId string, namespace string) *autoscaling.HorizontalPodAutoscaler {
 	hpa := autoscaling.HorizontalPodAutoscaler{
 		ObjectMeta: metav1.ObjectMeta{
@@ -521,131 +333,6 @@ func createIstioResources(mlDep *machinelearningv1alpha2.SeldonDeployment,
 	return vscs, drules
 }
 
-// Create istio virtual service and destination rule for explainer.
-// Explainers need one each with no traffic-splitting
-func createExplainerIstioResources(pSvcName string, p *machinelearningv1alpha2.PredictorSpec,
-	mlDep *machinelearningv1alpha2.SeldonDeployment,
-	seldonId string,
-	namespace string,
-	engine_http_port int,
-	engine_grpc_port int) ([]*istio.VirtualService, []*istio.DestinationRule) {
-
-	vsNameHttp := pSvcName + "-http"
-	if len(vsNameHttp) > 63 {
-		vsNameHttp = vsNameHttp[0:63]
-		vsNameHttp = strings.TrimSuffix(vsNameHttp, "-")
-	}
-
-	vsNameGrpc := pSvcName + "-grpc"
-	if len(vsNameGrpc) > 63 {
-		vsNameGrpc = vsNameGrpc[0:63]
-		vsNameGrpc = strings.TrimSuffix(vsNameGrpc, "-")
-	}
-
-	istio_gateway := getEnv(ENV_ISTIO_GATEWAY, "seldon-gateway")
-	httpVsvc := &istio.VirtualService{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      vsNameHttp,
-			Namespace: namespace,
-		},
-		Spec: istio.VirtualServiceSpec{
-			Hosts:    []string{"*"},
-			Gateways: []string{getAnnotation(mlDep, ANNOTATION_ISTIO_GATEWAY, istio_gateway)},
-			HTTP: []istio.HTTPRoute{
-				{
-					Match: []istio.HTTPMatchRequest{
-						{
-							URI: &v1alpha1.StringMatch{Prefix: "/seldon/" + namespace + "/" + mlDep.Name + "/" + p.Name + "/explainer/"},
-						},
-					},
-					Rewrite: &istio.HTTPRewrite{URI: "/"},
-				},
-			},
-		},
-	}
-
-	grpcVsvc := &istio.VirtualService{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      vsNameGrpc,
-			Namespace: namespace,
-		},
-		Spec: istio.VirtualServiceSpec{
-			Hosts:    []string{"*"},
-			Gateways: []string{getAnnotation(mlDep, ANNOTATION_ISTIO_GATEWAY, istio_gateway)},
-			HTTP: []istio.HTTPRoute{
-				{
-					Match: []istio.HTTPMatchRequest{
-						{
-							URI: &v1alpha1.StringMatch{Prefix: "/seldon.protos.Seldon/"},
-							Headers: map[string]v1alpha1.StringMatch{
-								"seldon":    v1alpha1.StringMatch{Exact: mlDep.Name}, //TODO: change this?
-								"namespace": v1alpha1.StringMatch{Exact: namespace},
-							},
-						},
-					},
-				},
-			},
-		},
-	}
-
-	routesHttp := make([]istio.HTTPRouteDestination, 1)
-	routesGrpc := make([]istio.HTTPRouteDestination, 1)
-	drules := make([]*istio.DestinationRule, 1)
-
-	drule := &istio.DestinationRule{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      pSvcName,
-			Namespace: namespace,
-		},
-		Spec: istio.DestinationRuleSpec{
-			Host: pSvcName,
-			Subsets: []istio.Subset{
-				{
-					Name: p.Name,
-					Labels: map[string]string{
-						"version": p.Labels["version"],
-					},
-				},
-			},
-		},
-	}
-
-	routesHttp[0] = istio.HTTPRouteDestination{
-		Destination: istio.Destination{
-			Host:   pSvcName,
-			Subset: p.Name,
-			Port: istio.PortSelector{
-				Number: uint32(engine_http_port),
-			},
-		},
-		Weight: int(100),
-	}
-	routesGrpc[0] = istio.HTTPRouteDestination{
-		Destination: istio.Destination{
-			Host:   pSvcName,
-			Subset: p.Name,
-			Port: istio.PortSelector{
-				Number: uint32(engine_grpc_port),
-			},
-		},
-		Weight: int(100),
-	}
-	drules[0] = drule
-
-	httpVsvc.Spec.HTTP[0].Route = routesHttp
-	grpcVsvc.Spec.HTTP[0].Route = routesGrpc
-	vscs := make([]*istio.VirtualService, 0, 2)
-	// explainer may not expose REST and grpc (presumably engine ensures predictors do?)
-	if engine_http_port > 0 {
-		vscs = append(vscs, httpVsvc)
-	}
-	if engine_grpc_port > 0 {
-		vscs = append(vscs, grpcVsvc)
-	}
-
-	return vscs, drules
-}
-
 func getEngineHttpPort() (engine_http_port int, err error) {
 	// Get engine http port from environment or use default
 	engine_http_port = DEFAULT_ENGINE_CONTAINER_PORT
@@ -774,35 +461,6 @@ func createComponents(r *ReconcileSeldonDeployment, mlDep *machinelearningv1alph
 		c.destinationRules = append(c.destinationRules, dstRule...)
 	}
 	return &c, nil
-}
-
-func addEngineToDeployment(mlDep *machinelearningv1alpha2.SeldonDeployment, p *machinelearningv1alpha2.PredictorSpec, engine_http_port int, engine_grpc_port int, pSvcName string, deploy *appsv1.Deployment) error {
-	engineContainer, err := createEngineContainer(mlDep, p, engine_http_port, engine_grpc_port)
-	if err != nil {
-		return err
-	}
-	deploy.Labels[machinelearningv1alpha2.Label_svc_orch] = "true"
-
-	deploy.Spec.Template.Spec.Containers = append(deploy.Spec.Template.Spec.Containers, *engineContainer)
-	//deploy.Spec.Template.Spec.ServiceAccountName = getEnv("ENGINE_CONTAINER_SERVICE_ACCOUNT_NAME", "seldon")
-	//deploy.Spec.Template.Spec.DeprecatedServiceAccount = deploy.Spec.Template.Spec.ServiceAccountName
-	//deploy.Spec.Template.Annotations = map[string]string{}
-	if deploy.Spec.Template.Annotations == nil {
-		deploy.Spec.Template.Annotations = make(map[string]string)
-	}
-	//overwrite annotations with predictor annotations
-	for _, ann := range p.Annotations {
-		deploy.Spec.Template.Annotations[ann] = p.Annotations[ann]
-	}
-	// Add prometheus annotations
-	deploy.Spec.Template.Annotations["prometheus.io/path"] = getEnv("ENGINE_PROMETHEUS_PATH", "/prometheus")
-	deploy.Spec.Template.Annotations["prometheus.io/port"] = strconv.Itoa(engine_http_port)
-	deploy.Spec.Template.Annotations["prometheus.io/scrape"] = "true"
-
-	deploy.ObjectMeta.Labels[machinelearningv1alpha2.Label_seldon_app] = pSvcName
-	deploy.Spec.Selector.MatchLabels[machinelearningv1alpha2.Label_seldon_app] = pSvcName
-	deploy.Spec.Template.ObjectMeta.Labels[machinelearningv1alpha2.Label_seldon_app] = pSvcName
-	return nil
 }
 
 //Creates Service for Predictor - exposed externally (ambassador or istio)
@@ -1045,141 +703,6 @@ func createDeploymentWithoutEngine(depName string, seldonId string, seldonPodSpe
 	}
 
 	return deploy
-}
-
-func createExplainer(r *ReconcileSeldonDeployment, mlDep *machinelearningv1alpha2.SeldonDeployment, p *machinelearningv1alpha2.PredictorSpec, c *components, pSvcName string) error {
-
-	if p.Explainer.Type != "" {
-
-		seldonId := machinelearningv1alpha2.GetSeldonDeploymentName(mlDep)
-
-		depName := machinelearningv1alpha2.GetExplainerDeploymentName(mlDep.ObjectMeta.Name, p)
-
-		explainerContainer := p.Explainer.ContainerSpec
-
-		if explainerContainer.Name == "" {
-			explainerContainer.Name = depName
-		}
-
-		if explainerContainer.ImagePullPolicy == "" {
-			explainerContainer.ImagePullPolicy = corev1.PullIfNotPresent
-		}
-
-		if p.Graph.Endpoint == nil {
-			p.Graph.Endpoint = &machinelearningv1alpha2.Endpoint{Type: machinelearningv1alpha2.REST}
-		}
-
-		if explainerContainer.Image == "" {
-			// TODO: should use explainer type but this is the only one available currently
-			explainerContainer.Image = "seldonio/alibiexplainer:0.1"
-		}
-
-		var portType string
-
-		// explainer can get port from spec or from containerSpec or fall back on default
-		var httpPort = 0
-		var grpcPort = 0
-		var portNum int32 = 9000
-		if p.Explainer.Endpoint != nil && p.Explainer.Endpoint.ServicePort != 0 {
-			portNum = p.Explainer.Endpoint.ServicePort
-		}
-		var pSvcEndpoint = ""
-		customPort := getPort(portType, explainerContainer.Ports)
-
-		if p.Explainer.Endpoint != nil && p.Explainer.Endpoint.Type == machinelearningv1alpha2.GRPC {
-			portType = "grpc"
-			grpcPort = int(portNum)
-			pSvcEndpoint = c.serviceDetails[pSvcName].GrpcEndpoint
-		} else {
-			portType = "http"
-			httpPort = int(portNum)
-			pSvcEndpoint = c.serviceDetails[pSvcName].HttpEndpoint
-		}
-
-		if customPort == nil {
-			explainerContainer.Ports = append(explainerContainer.Ports, corev1.ContainerPort{Name: portType, ContainerPort: portNum, Protocol: corev1.ProtocolTCP})
-		} else {
-			portNum = customPort.ContainerPort
-			portType = customPort.Name
-		}
-
-		if explainerContainer.LivenessProbe == nil {
-			explainerContainer.LivenessProbe = &corev1.Probe{Handler: corev1.Handler{TCPSocket: &corev1.TCPSocketAction{Port: intstr.FromString(portType)}}, InitialDelaySeconds: 60, PeriodSeconds: 5, SuccessThreshold: 1, FailureThreshold: 5, TimeoutSeconds: 1}
-		}
-		if explainerContainer.ReadinessProbe == nil {
-			explainerContainer.ReadinessProbe = &corev1.Probe{Handler: corev1.Handler{TCPSocket: &corev1.TCPSocketAction{Port: intstr.FromString(portType)}}, InitialDelaySeconds: 20, PeriodSeconds: 5, SuccessThreshold: 1, FailureThreshold: 7, TimeoutSeconds: 1}
-		}
-
-		// Add livecycle probe
-		if explainerContainer.Lifecycle == nil {
-			explainerContainer.Lifecycle = &corev1.Lifecycle{PreStop: &corev1.Handler{Exec: &corev1.ExecAction{Command: []string{"/bin/sh", "-c", "/bin/sleep 10"}}}}
-		}
-
-		explainerContainer.Args = []string{
-			"--explainerUri=/mnt/models",
-			"--explainer_name=" + mlDep.Name,
-			"--predict_url=" + "http://" + pSvcEndpoint + "/api/v0.1/predictions",
-			"--protocol=" + "seldon." + portType,
-			"--type=" + strings.ToLower(p.Explainer.Type),
-			"--http_port=" + strconv.Itoa(int(portNum))}
-
-		for k, v := range p.Explainer.Config {
-			//remote files in model location should get downloaded by initializer
-			if p.Explainer.ModelUri != "" {
-				v = strings.Replace(v, p.Explainer.ModelUri, "/mnt/models", 1)
-			}
-			arg := "--" + k + "=" + v
-			explainerContainer.Args = append(explainerContainer.Args, arg)
-		}
-		// see https://github.com/cliveseldon/kfserving/tree/explainer_update_jul/docs/samples/explanation/income for more
-
-		// Add Environment Variables - TODO: are these needed
-		if !utils.HasEnvVar(explainerContainer.Env, machinelearningv1alpha2.ENV_PREDICTIVE_UNIT_SERVICE_PORT) {
-			explainerContainer.Env = append(explainerContainer.Env, []corev1.EnvVar{
-				corev1.EnvVar{Name: machinelearningv1alpha2.ENV_PREDICTIVE_UNIT_SERVICE_PORT, Value: strconv.Itoa(int(portNum))},
-				corev1.EnvVar{Name: machinelearningv1alpha2.ENV_PREDICTIVE_UNIT_ID, Value: explainerContainer.Name},
-				corev1.EnvVar{Name: machinelearningv1alpha2.ENV_PREDICTOR_ID, Value: p.Name},
-				corev1.EnvVar{Name: machinelearningv1alpha2.ENV_SELDON_DEPLOYMENT_ID, Value: mlDep.ObjectMeta.Name},
-			}...)
-		}
-
-		seldonPodSpec := machinelearningv1alpha2.SeldonPodSpec{Spec: corev1.PodSpec{
-			Containers: []corev1.Container{explainerContainer},
-		}}
-		deploy := createDeploymentWithoutEngine(depName, seldonId, &seldonPodSpec, p, mlDep)
-
-		deploy, err := InjectModelInitializer(deploy, explainerContainer.Name, p.Explainer.ModelUri, p.Explainer.ServiceAccountName, r.Client)
-
-		// for explainer use same service name as its Deployment
-		eSvcName := machinelearningv1alpha2.GetExplainerDeploymentName(mlDep.ObjectMeta.Name, p)
-
-		deploy.ObjectMeta.Labels[machinelearningv1alpha2.Label_seldon_app] = eSvcName
-		deploy.Spec.Template.ObjectMeta.Labels[machinelearningv1alpha2.Label_seldon_app] = eSvcName
-
-		c.deployments = append(c.deployments, deploy)
-
-		//Create Service to be exposed externally (ambassador or istio)
-		eSvc, err := createPredictorService(eSvcName, seldonId, p, mlDep, httpPort, grpcPort)
-		if err != nil {
-			return err
-		}
-		c.services = append(c.services, eSvc)
-		c.serviceDetails[eSvcName] = &machinelearningv1alpha2.ServiceStatus{
-			SvcName:      eSvcName,
-			HttpEndpoint: eSvcName + "." + eSvc.Namespace + ":" + strconv.Itoa(httpPort),
-			ExplainerFor: machinelearningv1alpha2.GetPredictorKey(mlDep, p),
-		}
-		if grpcPort > 0 {
-			c.serviceDetails[eSvcName].GrpcEndpoint = eSvcName + "." + eSvc.Namespace + ":" + strconv.Itoa(grpcPort)
-		}
-		if getEnv(ENV_ISTIO_ENABLED, "false") == "true" {
-			vsvcs, dstRule := createExplainerIstioResources(eSvcName, p, mlDep, seldonId, getNamespace(mlDep), httpPort, grpcPort)
-			c.virtualServices = append(c.virtualServices, vsvcs...)
-			c.destinationRules = append(c.destinationRules, dstRule...)
-		}
-	}
-
-	return nil
 }
 
 func getPort(name string, ports []corev1.ContainerPort) *corev1.ContainerPort {
